add top-level Cargo.toml workspace configuration file
This also moves common configuration from each package's Cargo.toml file to the workspace file. And it makes minor adjustments to the names of the packages for consistency with the names of their directories. Finally, it comments out some calls to the sensor API that was commented out in a previous commit on this branch.